--- models.py
@@ -1,19 +1,0 @@
-from app import db
-
-__all__ = ['UserModel']
-
-class UserModel(db.Model):
-    __tablename__ = 'users2'
-
-    id = db.Column(db.Integer, primary_key = True)
-    username = db.Column(db.String(120), unique = True, nullable = False)
-    password = db.Column(db.String(120), nullable = False)
-    email = db.Column(db.String(120), nullable = False)
-    
-    def save_to_db(self):
-        db.session.add(self)
-        db.session.commit()
-    
-    @classmethod
-    def find_by_username(cls, username):
-        return cls.query.filter_by(username = username).first()
